split change feed tests
